game service skeleton

+++ src/main/kotlin/ru/afanasyev/rocketapi/domain/ObjectPosition.kt
@@ -1,3 +1,3 @@
 package ru.afanasyev.rocketapi.domain
 
-data class ObjectPosition(val x: Long, val y: Long)
+data class ObjectPosition(val x: Double, val y: Double)
